bpiotrowski adopted and moved the go package from [community] to [extra], against the will of the (then) two co-maintainers of the go package.
He then proceeded to remove the list of maintainers and contributors from the top of the PKGBUILD.

The package was also split into "go" and "go-pie".

Now that the package is back in [community], I'm adding back the list of maintainers and contributors.

Also updated the URL to https:// instead of http://, combined two "rm" commands to a single command and added the missing license file to the go-pie package.